Quiet and fix a lot of 64-to-32-bit warnings. More properties added.

+++ Seashore/source/application/SeaProxy.m
@@ -142,7 +142,7 @@
 
 - (IBAction)setColorSpace:(id)sender
 {
-	[[gCurrentDocument contents] convertToType:[sender tag] - 240];
+	[[gCurrentDocument contents] convertToType:(int)([sender tag] - 240)];
 }
 
 - (IBAction)toggleLinked:(id)sender
